+++ IronNet/src/ironattack/session_hijack.h
@@ -1,0 +1,6 @@
+#ifndef IRON_SESSION_HIJACK_H
+#define IRON_SESSION_HIJACK_H
+
+int cmd_session_hijack(int argc, char **argv);
+
+#endif